requirements/common.txt
alembic==1.7.7
bleach==5.0.0
blinker==1.4
Bootstrap_Flask==2.0.2
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email_validator==1.2.0
Flask==2.1.2
Flask-CKEditor==0.4.6
Flask-HTTPAuth==4.6.0
Flask_Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask_Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask_WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
Pillow==9.1.0
PyJWT==2.3.0
setuptools==58.1.0
six==1.16.0
SQLAlchemy==1.4.36
webencodings==0.5.1
Werkzeug==2.1.2
WTForms==3.0.1
